+++ packages/framer-motion/cypress/integration/while-in-view.ts
@@ -139,28 +139,4 @@ describe("whileInView", () => {
                 expect($element.innerHTML).to.equal("In")
             })
     })
-
-    it.skip("If IntersectionObserver doesn't exist, immediately animates to whileInView", () => {
-        cy.visit("?test=while-in-view&delete=true")
-            .wait(100)
-            .get("#box")
-            .should(([$element]: any) => {
-                expect($element.style.backgroundColor).to.equal(
-                    "rgb(0, 255, 0)"
-                )
-                expect($element.innerHTML).to.equal("In")
-            })
-    })
-
-    it.skip("If IntersectionObserver doesn't exist, and fallback is false, don't animate to whileInView", () => {
-        cy.visit("?test=while-in-view&delete=true&disableFallback=true")
-            .wait(100)
-            .get("#box")
-            .should(([$element]: any) => {
-                expect($element.style.backgroundColor).not.to.equal(
-                    "rgb(0, 255, 0)"
-                )
-                expect($element.innerHTML).not.to.equal("In")
-            })
-    })
 })
