+++ packages/framer-motion/src/events/add-dom-event.ts
@@ -1,0 +1,10 @@
+export function addDomEvent(
+    target: EventTarget,
+    eventName: string,
+    handler: EventListener,
+    options: AddEventListenerOptions = { passive: true }
+) {
+    target.addEventListener(eventName, handler, options)
+
+    return () => target.removeEventListener(eventName, handler)
+}
