+++ packages/framer-motion/src/gestures/pan/PanSession.ts
@@ -2,7 +2,7 @@ import { EventInfo } from "../../events/types"
 import { extractEventInfo } from "../../events/event-info"
 import { sync, cancelSync } from "../../frameloop"
 import { secondsToMilliseconds } from "../../utils/time-conversion"
-import { addPointerEvent } from "../../events/use-pointer-event"
+import { addPointerEvent } from "../../events/add-pointer-event"
 import { Point, TransformPoint } from "../../projection/geometry/types"
 import { pipe } from "../../utils/pipe"
 import { distance2D } from "../../utils/distance"
